+++ packages/platform-browser/src/IndexedDbMigration.ts
@@ -1,14 +1,11 @@
 /**
  * @since 1.0.0
  */
-import { TypeIdError } from "@effect/platform/Error"
-import * as Cause from "effect/Cause"
-import * as Effect from "effect/Effect"
-import * as HashMap from "effect/HashMap"
-import { type Pipeable, pipeArguments } from "effect/Pipeable"
-import * as Schema from "effect/Schema"
+import type * as Effect from "effect/Effect"
+import { type Pipeable } from "effect/Pipeable"
 import type * as IndexedDbTable from "./IndexedDbTable.js"
 import type * as IndexedDbVersion from "./IndexedDbVersion.js"
+import * as internal from "./internal/indexedDbMigration.js"
 
 type IsStringLiteral<T> = T extends string ? string extends T ? false
   : true
@@ -45,9 +42,7 @@ export type IndexFromTable<
  * @since 1.0.0
  * @category type ids
  */
-export const TypeId: unique symbol = Symbol.for(
-  "@effect/platform-browser/IndexedDbMigration"
-)
+export const TypeId: unique symbol = internal.TypeId
 
 /**
  * @since 1.0.0
@@ -59,9 +54,7 @@ export type TypeId = typeof TypeId
  * @since 1.0.0
  * @category type ids
  */
-export const ErrorTypeId: unique symbol = Symbol.for(
-  "@effect/platform-browser/IndexedDbMigration/IndexedDbMigrationError"
-)
+export const ErrorTypeId: unique symbol = internal.ErrorTypeId
 
 /**
  * @since 1.0.0
@@ -73,322 +66,7 @@ export type ErrorTypeId = typeof ErrorTypeId
  * @since 1.0.0
  * @category errors
  */
-export class IndexedDbMigrationError extends TypeIdError(
-  ErrorTypeId,
-  "IndexedDbMigrationError"
-)<{
-  readonly reason:
-    | "OpenError"
-    | "TransactionError"
-    | "DecodeError"
-    | "Blocked"
-    | "UpgradeError"
-    | "MissingTable"
-    | "MissingIndex"
-  readonly cause: unknown
-}> {
-  get message() {
-    return this.reason
-  }
-}
-
-export interface MigrationApi<
-  Source extends IndexedDbVersion.IndexedDbVersion.AnyWithProps
-> {
-  readonly createObjectStore: <
-    A extends IndexedDbTable.IndexedDbTable.TableName<
-      IndexedDbVersion.IndexedDbVersion.Tables<Source>
-    >
-  >(table: A) => Effect.Effect<globalThis.IDBObjectStore, IndexedDbMigrationError>
-
-  readonly deleteObjectStore: <
-    A extends IndexedDbTable.IndexedDbTable.TableName<
-      IndexedDbVersion.IndexedDbVersion.Tables<Source>
-    >
-  >(table: A) => Effect.Effect<void, IndexedDbMigrationError>
-
-  readonly createIndex: <
-    A extends IndexedDbTable.IndexedDbTable.TableName<
-      IndexedDbVersion.IndexedDbVersion.Tables<Source>
-    >
-  >(
-    table: A,
-    indexName: IndexFromTable<Source, A>,
-    options?: IDBIndexParameters
-  ) => Effect.Effect<globalThis.IDBIndex, IndexedDbMigrationError>
-
-  readonly deleteIndex: <
-    A extends IndexedDbTable.IndexedDbTable.TableName<
-      IndexedDbVersion.IndexedDbVersion.Tables<Source>
-    >
-  >(table: A, indexName: IndexFromTable<Source, A>) => Effect.Effect<void, IndexedDbMigrationError>
-
-  readonly getAll: <
-    A extends IndexedDbTable.IndexedDbTable.TableName<
-      IndexedDbVersion.IndexedDbVersion.Tables<Source>
-    >
-  >(table: A) => Effect.Effect<
-    Array<
-      Schema.Schema.Type<
-        IndexedDbTable.IndexedDbTable.TableSchema<
-          IndexedDbTable.IndexedDbTable.WithName<
-            IndexedDbVersion.IndexedDbVersion.Tables<Source>,
-            A
-          >
-        >
-      >
-    >,
-    IndexedDbMigrationError
-  >
-
-  readonly insert: <
-    A extends IndexedDbTable.IndexedDbTable.TableName<
-      IndexedDbVersion.IndexedDbVersion.Tables<Source>
-    >
-  >(
-    table: A,
-    data: Schema.Schema.Encoded<
-      IndexedDbTable.IndexedDbTable.TableSchema<
-        IndexedDbTable.IndexedDbTable.WithName<
-          IndexedDbVersion.IndexedDbVersion.Tables<Source>,
-          A
-        >
-      >
-    >
-  ) => Effect.Effect<globalThis.IDBValidKey, IndexedDbMigrationError>
-
-  readonly insertAll: <
-    A extends IndexedDbTable.IndexedDbTable.TableName<
-      IndexedDbVersion.IndexedDbVersion.Tables<Source>
-    >
-  >(
-    table: A,
-    dataList: ReadonlyArray<
-      Schema.Schema.Encoded<
-        IndexedDbTable.IndexedDbTable.TableSchema<
-          IndexedDbTable.IndexedDbTable.WithName<
-            IndexedDbVersion.IndexedDbVersion.Tables<Source>,
-            A
-          >
-        >
-      >
-    >
-  ) => Effect.Effect<globalThis.IDBValidKey, IndexedDbMigrationError>
-}
-
-/** @internal */
-export const migrationApi = <
-  Source extends IndexedDbVersion.IndexedDbVersion.AnyWithProps
->(
-  database: IDBDatabase,
-  transaction: IDBTransaction,
-  source: Source
-): MigrationApi<Source> => {
-  const insert = <
-    A extends IndexedDbTable.IndexedDbTable.TableName<
-      IndexedDbVersion.IndexedDbVersion.Tables<Source>
-    >
-  >(
-    table: A,
-    data: Schema.Schema.Encoded<
-      IndexedDbTable.IndexedDbTable.TableSchema<
-        IndexedDbTable.IndexedDbTable.WithName<
-          IndexedDbVersion.IndexedDbVersion.Tables<Source>,
-          A
-        >
-      >
-    >
-  ) =>
-    Effect.gen(function*() {
-      const { tableSchema } = yield* HashMap.get(source.tables, table).pipe(
-        Effect.catchTag(
-          "NoSuchElementException",
-          (cause) =>
-            new IndexedDbMigrationError({
-              reason: "MissingTable",
-              cause
-            })
-        )
-      )
-
-      yield* Schema.decodeUnknown(tableSchema)(data).pipe(
-        Effect.catchTag("ParseError", (cause) =>
-          new IndexedDbMigrationError({
-            reason: "DecodeError",
-            cause
-          }))
-      )
-
-      return yield* Effect.async<globalThis.IDBValidKey, IndexedDbMigrationError>((resume) => {
-        const objectStore = transaction.objectStore(table)
-        const request = objectStore.add(data)
-
-        request.onerror = () => {
-          resume(
-            Effect.fail(
-              new IndexedDbMigrationError({
-                reason: "TransactionError",
-                cause: request.error
-              })
-            )
-          )
-        }
-
-        request.onsuccess = () => {
-          resume(Effect.succeed(request.result))
-        }
-      })
-    })
-
-  return {
-    insert,
-    insertAll: (table, dataList) => Effect.all(dataList.map((data) => insert(table, data))),
-
-    createObjectStore: (table) =>
-      Effect.gen(function*() {
-        const createTable = yield* HashMap.get(source.tables, table).pipe(
-          Effect.catchTag(
-            "NoSuchElementException",
-            (cause) =>
-              new IndexedDbMigrationError({
-                reason: "MissingTable",
-                cause
-              })
-          )
-        )
-
-        return yield* Effect.try({
-          try: () =>
-            database.createObjectStore(
-              createTable.tableName,
-              createTable.options
-            ),
-          catch: (cause) =>
-            new IndexedDbMigrationError({
-              reason: "TransactionError",
-              cause
-            })
-        })
-      }),
-
-    deleteObjectStore: (table) =>
-      Effect.gen(function*() {
-        const createTable = yield* HashMap.get(source.tables, table).pipe(
-          Effect.catchTag(
-            "NoSuchElementException",
-            (cause) =>
-              new IndexedDbMigrationError({
-                reason: "MissingTable",
-                cause
-              })
-          )
-        )
-
-        return yield* Effect.try({
-          try: () => database.deleteObjectStore(createTable.tableName),
-          catch: (cause) =>
-            new IndexedDbMigrationError({
-              reason: "TransactionError",
-              cause
-            })
-        })
-      }),
-
-    createIndex: (table, indexName, options) =>
-      Effect.gen(function*() {
-        const store = transaction.objectStore(table)
-        const sourceTable = HashMap.unsafeGet(source.tables, table)
-
-        const keyPath = yield* Effect.fromNullable(
-          sourceTable.options?.indexes[indexName] ?? undefined
-        ).pipe(
-          Effect.catchTag("NoSuchElementException", (error) =>
-            new IndexedDbMigrationError({
-              reason: "MissingIndex",
-              cause: Cause.fail(error)
-            }))
-        )
-
-        return yield* Effect.try({
-          try: () => store.createIndex(indexName, keyPath, options),
-          catch: (cause) =>
-            new IndexedDbMigrationError({
-              reason: "TransactionError",
-              cause
-            })
-        })
-      }),
-
-    deleteIndex: (table, indexName) =>
-      Effect.gen(function*() {
-        const store = transaction.objectStore(table)
-        return yield* Effect.try({
-          try: () => store.deleteIndex(indexName),
-          catch: (cause) =>
-            new IndexedDbMigrationError({
-              reason: "TransactionError",
-              cause
-            })
-        })
-      }),
-
-    getAll: (table) =>
-      Effect.gen(function*() {
-        const { tableName, tableSchema } = yield* HashMap.get(
-          source.tables,
-          table
-        ).pipe(
-          Effect.catchTag(
-            "NoSuchElementException",
-            (cause) =>
-              new IndexedDbMigrationError({
-                reason: "MissingTable",
-                cause
-              })
-          )
-        )
-
-        const data = yield* Effect.async<any, IndexedDbMigrationError>(
-          (resume) => {
-            const store = transaction.objectStore(tableName)
-            const request = store.getAll()
-
-            request.onerror = () => {
-              resume(
-                Effect.fail(
-                  new IndexedDbMigrationError({
-                    reason: "TransactionError",
-                    cause: request.error
-                  })
-                )
-              )
-            }
-
-            request.onsuccess = () => {
-              resume(Effect.succeed(request.result))
-            }
-          }
-        )
-
-        const tableSchemaArray = Schema.Array(
-          tableSchema
-        ) as unknown as IndexedDbTable.IndexedDbTable.TableSchema<
-          IndexedDbTable.IndexedDbTable.WithName<
-            IndexedDbVersion.IndexedDbVersion.Tables<Source>,
-            typeof tableName
-          >
-        >
-
-        return yield* Schema.decodeUnknown(tableSchemaArray)(data).pipe(
-          Effect.catchTag("ParseError", (cause) =>
-            new IndexedDbMigrationError({
-              reason: "DecodeError",
-              cause
-            }))
-        )
-      })
-  }
-}
+export const IndexedDbMigrationError = internal.IndexedDbMigrationError
 
 /**
  * @since 1.0.0
@@ -408,7 +86,7 @@ export declare namespace IndexedDbMigration {
     readonly [TypeId]: TypeId
     readonly _tag: "Initial"
     readonly version: InitialVersion
-    readonly execute: (toQuery: MigrationApi<InitialVersion>) => Effect.Effect<void, Error>
+    readonly execute: (toQuery: internal.MigrationApi<InitialVersion>) => Effect.Effect<void, Error>
 
     readonly add: <
       Version extends IndexedDbVersion.IndexedDbVersion.AnyWithProps,
@@ -416,8 +94,8 @@ export declare namespace IndexedDbMigration {
     >(
       version: Version,
       execute: (
-        fromQuery: MigrationApi<InitialVersion>,
-        toQuery: MigrationApi<Version>
+        fromQuery: internal.MigrationApi<InitialVersion>,
+        toQuery: internal.MigrationApi<Version>
       ) => Effect.Effect<void, MigrationError>
     ) => Migration<InitialVersion, Version, MigrationError | Error>
   }
@@ -439,8 +117,8 @@ export declare namespace IndexedDbMigration {
     readonly fromVersion: FromVersion
     readonly toVersion: ToVersion
     readonly execute: (
-      fromQuery: MigrationApi<FromVersion>,
-      toQuery: MigrationApi<ToVersion>
+      fromQuery: internal.MigrationApi<FromVersion>,
+      toQuery: internal.MigrationApi<ToVersion>
     ) => Effect.Effect<void, Error>
   }
 
@@ -470,74 +148,8 @@ export declare namespace IndexedDbMigration {
   export type AnyInitial = Initial<IndexedDbVersion.IndexedDbVersion.AnyWithProps, any>
 }
 
-const Proto = {
-  [TypeId]: TypeId,
-  pipe() {
-    return pipeArguments(this, arguments)
-  }
-}
-
-const makeInitialProto = <
-  InitialVersion extends IndexedDbVersion.IndexedDbVersion.AnyWithProps,
-  Error
->(
-  initialVersion: InitialVersion,
-  init: (
-    toQuery: MigrationApi<InitialVersion>
-  ) => Effect.Effect<void, Error>
-): IndexedDbMigration.Initial<InitialVersion, Error> => {
-  function IndexedDbMigration() {}
-  Object.setPrototypeOf(IndexedDbMigration, Proto)
-  IndexedDbMigration.version = initialVersion
-  IndexedDbMigration.execute = init
-  IndexedDbMigration._tag = "Initial"
-
-  IndexedDbMigration.add = <
-    Version extends IndexedDbVersion.IndexedDbVersion.AnyWithProps
-  >(
-    version: Version,
-    execute: (
-      fromQuery: MigrationApi<InitialVersion>,
-      toQuery: MigrationApi<Version>
-    ) => Effect.Effect<void, Error>
-  ) =>
-    makeProto({
-      fromVersion: initialVersion,
-      toVersion: version,
-      execute,
-      previous: IndexedDbMigration as any
-    })
-
-  return IndexedDbMigration as any
-}
-
-const makeProto = <
-  FromVersion extends IndexedDbVersion.IndexedDbVersion.AnyWithProps,
-  ToVersion extends IndexedDbVersion.IndexedDbVersion.AnyWithProps,
-  Error
->(options: {
-  readonly previous:
-    | IndexedDbMigration.Migration<FromVersion, ToVersion, Error>
-    | IndexedDbMigration.Initial<FromVersion, Error>
-  readonly fromVersion: FromVersion
-  readonly toVersion: ToVersion
-  readonly execute: (
-    fromQuery: MigrationApi<FromVersion>,
-    toQuery: MigrationApi<ToVersion>
-  ) => Effect.Effect<void, Error>
-}): IndexedDbMigration.Migration<FromVersion, ToVersion, Error> => {
-  function IndexedDbMigration() {}
-  Object.setPrototypeOf(IndexedDbMigration, Proto)
-  IndexedDbMigration.previous = options.previous
-  IndexedDbMigration.fromVersion = options.fromVersion
-  IndexedDbMigration.toVersion = options.toVersion
-  IndexedDbMigration.execute = options.execute
-  IndexedDbMigration._tag = "Migration"
-  return IndexedDbMigration as any
-}
-
 /**
  * @since 1.0.0
  * @category constructors
  */
-export const make = makeInitialProto
+export const make = internal.makeInitialProto
